% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_function.R
\name{dog_function}
\alias{dog_function}
\title{A Dog Function

This function investigates your love of dogs.}
\usage{
dog_function(love = TRUE)
}
\arguments{
\item{love}{Do you love dogs? Defaults to TRUE.}
}
\description{
A Dog Function

This function investigates your love of dogs.
}
\examples{
dog_function()
}
\keyword{dogs}
