% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_study.R
\docType{data}
\name{dog_study}
\alias{dog_study}
\title{Dog parameters}
\format{
An object of class \code{"tibble"} \code `dog_id`
}
\source{
\href{https://i.redd.it/fd0f3p43mqqy.jpg}{Archive}
}
\usage{
data(dog_study)
}
\description{
Dataset containing dog parameters gathered during from the (infamous) dog improvement program 2020 - 2022.
}
\examples{
data(dog_study)
summary(dog_study)
}
\references{
L'bardo et al. (2022) Dog Genetics 231:177-216
(\href{https://external-preview.redd.it/3IKvKF2xhJOwsggp_N3HwklJogvj4vG0DgdHircthj4.jpg?auto=webp&s=18c64cd8eab1e12458d1c56b620bf0452639260e}{PubMed})
}
\keyword{datasets}
