% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogs-package.R
\docType{package}
\name{dogs-package}
\alias{dogs}
\alias{dogs-package}
\title{dogs: Functions about dogs}
\description{
This package provides a set of functions that illustrate how dogs truly are man's best friends
}
\author{
\strong{Maintainer}: Henry Wang \email{henrywangnl@gmail.com}

}
\keyword{internal}
