% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{helloworld}
\alias{helloworld}
\title{Hello world function}
\usage{
helloworld(name = "my friend")
}
\arguments{
\item{name}{say hello to the person. Defaults to 'my friend'.}
}
\description{
This function allows you to say hello to everyone.
}
\examples{
helloworld(name='David')
}
\keyword{hello}
