% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Impute.R
\name{impute}
\alias{impute}
\title{Imputation Function}
\usage{
impute(A = 0, loged = FALSE, prepared = FALSE, method = "NMF")
}
\arguments{
\item{a}{data matrix A, which represent gene expressions for each cell/sample. rows = genes and cols= cells, first column should be gene names, first row should be cell names.}

\item{b}{loged should be TRUE if your data is loged, default is FALSE}

\item{c}{prepared to save time prepared should be set to TRUE if you have ran impute on your data before, and you just want to try different methods, default is FALSE}

\item{d}{method could be set to "NMF", "KNN" or "SlopeOne". default is "NMF"}
}
\description{
This function will Impute zero values of your data.
}
\examples{
impute(A)
impute(A, method="NMF" )
}
\keyword{imputation}
