% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_functions.R
\name{jump}
\alias{jump}
\title{Jump}
\usage{
jump(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The mean of the vector.
}
\description{
This dog doesn't jump.
}
\examples{
jump(c(0.5, 0.2, 0.3))
}
