% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dog_angle.R
\name{plot_dog_angle}
\alias{plot_dog_angle}
\title{Plot Dog Angles

Plot dog angles and interactions with carbon footpring and dog criminal record}
\usage{
plot_dog_angle(dog_study)
}
\arguments{
\item{dog_study}{dog_study data set included with the package.}
}
\value{
\code{gg} ggplot object
}
\description{
Plot Dog Angles

Plot dog angles and interactions with carbon footpring and dog criminal record
}
\examples{
data(dog_study)
dog_function(dog_study)
}
\keyword{dogs}
