% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dog_function.R
\name{dog_function}
\alias{dog_function}
\title{A Dog Function}
\usage{
dog_function(love = TRUE)
}
\arguments{
\item{love}{Do you love dogs? Defaults to TRUE}
}
\description{
(Based of Hilary Parker's Cat Function - https://hilaryparker.com/2014/04/29/writing-an-r-package-from-scratch/)
This function allows you to express your love of dogs
}
\examples{
dog_function()
}
\keyword{dogs}
