% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogs-package.R
\docType{package}
\name{dogs-package}
\alias{dogs}
\alias{dogs-package}
\title{dogs: Functions about dogs}
\description{
This packages provides a set of functions that illustrate how dogs truly are man's best friend.
}
\author{
\strong{Maintainer}: Tomas Westlake \email{tom.westlake121@gmail.com}

}
\keyword{internal}
