% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogs-package.R
\docType{package}
\name{dogs-package}
\alias{dogs}
\alias{dogs-package}
\title{dogs: functions about dogs}
\description{
This package provides a function about dogs.
}
\author{
\strong{Maintainer}: Jimmy Glenn \email{jimmykglenn@gmail.com}

}
\keyword{internal}
