% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dogs-package.R
\docType{package}
\name{dogs-package}
\alias{dogs}
\alias{dogs-package}
\title{dogs: Dogs Are Better Than Cats}
\description{
This is for playing with.
}
\author{
\strong{Maintainer}: Ken Butler \email{nxskok@gmail.com}

}
\keyword{internal}
